requirements.txt
python-telegram-bot==20.8
pytz~=2024.1